# llama : refactor quantization to avoid <mutex> header