ref https://github.com/ggerganov/llama.cpp/pull/4613#issuecomment-1872320224