@cebtenzzre 2024-01-03
The comment suggests this change should increase performance, but a quick test I ran shows a decrease. This lock probably doesn't have much contention.

7B source on ramfs, dest is /dev/null
|model|quant|master (ms)|PR (ms)|speedup|
|-|-|-|-|-|
|7B|Q4_0|4880.14±69.10|5978.96±69.61|0.82|