@ggerganov 2024-01-04
Thanks for checking - I did tests just on macOS and didn't observe any slowdown. Will think about this some more - I might still decide to merge it because it reduces the C++ surface and in my experience, having less synchronization primitives is always better